Decreased length of possible recipes to speed up algorithm
I adjusted the maximum permutation length when checking for possible recipes. When working with the larger recipe datasets, this saves a large amount of time.
I found that for most ingredients above 3 words, it is usually giving different options for the single ingredient (e.g. walnuts, almonds, cashews, peanuts, pistachios, pecans, or hazelnuts) which just adds a lot of computation time as the program won't recognize the ingredient until it gets down to permutations of size 1 (it starts with the longest possible ones first).