Added cosine similarity functionality
I finished the cosine similarity algorithm based on the basic recipe ingredients. I plan to add in more functionality if I have extra time or just focus on implementing a different algorithm altogether.

I learned that calculating a cosine similarity for all recipes ends up taking a large amount of space, so I limited it to only take in a file with 15k recipes.