Cleaned up cosFoodRec.py and added in the last functionality I had planned.
In cosFoodRec.py:
I added in two simple CLA's, one required for the dataset file, and the other a flag for using the recipe names in the recommendation algorithm. This can make it recommend recipes that have a similar title instead of basing it solely on ingredients.
I also cleaned up main and added in a few functions for readability.

I believe I am finished with cosFoodRec.py. If there are any more features I would like to implement they would probably be minor. I will now focus on either creating a simple web application for this program or I will change it up and make it work with another recommendation algorithm.